requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
pydantic==2.5.0